requirements.txt
asgiref==3.2.5
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
Flask==1.1.1
google-api-python-client==1.7.12
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-decouple==3.3
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
whitenoise==5.0.1
